Vectors of Univariate Sample Statistics

Descriptives(x) gives a vector of sample statistics for use in codebook.

Usage

Descriptives(x,...)
# S4 method for class 'atomic'
Descriptives(x, weights = NULL, ...)
# S4 method for class 'item.vector'
Descriptives(x, weights = NULL, ...)

Arguments

x

an atomic vector or "item.vector" object.

weights

an optional vector of weights.

...

further arguments, to be passed to future methods.

Value

A numeric vector of sample statistics, containing the range, the mean, the standard deviation, the skewness and the (excess) kurtosis.

Examples

x <- rnorm(100)
Descriptives(x)
#>         Min         Max        Mean    Std.Dev.    Skewness    Kurtosis 
#> -2.61233433  2.75541758  0.07051001  1.03962126 -0.19009035  0.35012963