Data Set Objects

"data.set" objects are collections of "item" objects, with similar semantics as data frames. They are distinguished from data frames so that coercion by as.data.fame leads to a data frame that contains only vectors and factors. Nevertheless most methods for data frames are inherited by data sets, except for the method for the within generic function. For the within method for data sets, see the details section.

Thus data preparation using data sets retains all informations about item annotations, labels, missing values etc. While (mostly automatic) conversion of data sets into data frames makes the data amenable for the use of R's statistical functions.

dsView is a function that displays data sets in a similar manner as View displays data frames. (View works with data sets as well, but changes them first into data frames.)

Usage

data.set(...,row.names = NULL, check.rows = FALSE, check.names = TRUE,
    stringsAsFactors = FALSE, document = NULL)
as.data.set(x, row.names=NULL, ...)
# S4 method for list
as.data.set(x,row.names=NULL,...)
is.data.set(x)
# S3 method for data.set
as.data.frame(x, row.names = NULL, optional = FALSE, ...)
# S4 method for data.set
within(data, expr, ...)

dsView(x)

# S4 method for data.set
head(x,n=20,...)
# S4 method for data.set
tail(x,n=20,...)

Arguments

...

For the data.set function several vectors or items, for within further, ignored arguments.

row.names, check.rows, check.names, stringsAsFactors, optional

arguments as in data.frame or as.data.frame, respectively.

document

NULL or an optional character vector that contains documenation of the data.

x

for is.data.set(x), any object; for as.data.frame(x,...) and dsView(x) a "data.set" object.

data

a data set, that is, an object of class "data.set".

expr

an expression, or several expressions enclosed in curly braces.

n

integer; the number of rows to be shown by head or tail

Details

The as.data.frame method for data sets is just a copy of the method for list. Consequently, all items in the data set are coerced in accordance to their measurement setting, see as.vector,item-method and measurement.

The within method for data sets has the same effect as the within method for data frames, apart from two differences: all results of the computations are coerced into items if they have the appropriate length, otherwise, they are automatically dropped.

Currently only one method for the generic function as.data.set is defined: a method for "importer" objects.

Value

data.set and the within method for data sets returns a "data.set" object, is.data.set

returns a logical value, and as.data.frame returns a data frame.

Examples

Data <- data.set(
          vote = sample(c(1,2,3,8,9,97,99),size=300,replace=TRUE),
          region = sample(c(rep(1,3),rep(2,2),3,99),size=300,replace=TRUE),
          income = exp(rnorm(300,sd=.7))*2000
          )

Data <- within(Data,{
  description(vote) <- "Vote intention"
  description(region) <- "Region of residence"
  description(income) <- "Household income"
  wording(vote) <- "If a general election would take place next tuesday,
                    the candidate of which party would you vote for?"
  wording(income) <- "All things taken into account, how much do all
                    household members earn in sum?"
  foreach(x=c(vote,region),{
    measurement(x) <- "nominal"
    })
  measurement(income) <- "ratio"
  labels(vote) <- c(
                    Conservatives         =  1,
                    Labour                =  2,
                    "Liberal Democrats"   =  3,
                    "Don't know"          =  8,
                    "Answer refused"      =  9,
                    "Not applicable"      = 97,
                    "Not asked in survey" = 99)
  labels(region) <- c(
                    England               =  1,
                    Scotland              =  2,
                    Wales                 =  3,
                    "Not applicable"      = 97,
                    "Not asked in survey" = 99)
  foreach(x=c(vote,region,income),{
    annotation(x)["Remark"] <- "This is not a real survey item, of course ..."
    })
  missing.values(vote) <- c(8,9,97,99)
  missing.values(region) <- c(97,99)

  # These to variables do not appear in the
  # the resulting data set, since they have the wrong length.
  junk1 <- 1:5
  junk2 <- matrix(5,4,4)
  
})
#> Warning: Variables 'junk1','junk2' have wrong length, removing them.
# Since data sets may be huge, only a
# part of them are 'show'n
Data
#> 
#> Data set with 300 observations and 3 variables
#> 
#>                    vote               region    income
#>  1 *Not asked in survey             Scotland  431.9836
#>  2 *Not asked in survey             Scotland 3076.6872
#>  3      *Not applicable *Not asked in survey  887.9832
#>  4          *Don't know *Not asked in survey 1416.5029
#>  5               Labour              England 2445.3780
#>  6      *Not applicable *Not asked in survey 2022.1012
#>  7      *Answer refused              England  950.5021
#>  8        Conservatives             Scotland  684.3520
#>  9          *Don't know             Scotland 1567.1689
#> 10          *Don't know             Scotland 1334.4476
#> 11               Labour                Wales 3728.2146
#> 12 *Not asked in survey *Not asked in survey 4586.1085
#> 13      *Not applicable              England 1202.7189
#> 14    Liberal Democrats              England 3632.0289
#> 15          *Don't know             Scotland 6660.9266
#> 16    Liberal Democrats *Not asked in survey  701.1090
#> 17               Labour                Wales 3005.1291
#> 18      *Answer refused                Wales 2812.0180
#> 19 *Not asked in survey              England 4548.7792
#> 20      *Answer refused             Scotland  967.8923
#> 21          *Don't know              England 2097.1810
#> 22               Labour *Not asked in survey  530.8235
#> 23          *Don't know             Scotland  803.4231
#> 24          *Don't know *Not asked in survey  807.8204
#> 25      *Not applicable              England  655.0824
#> .. .................... .................... .........
#> (25 of 300 observations shown)

if (FALSE) {

# If we insist on seeing all, we can use 'print' instead
print(Data)
}

str(Data)
#> Data set with 300 obs. of 3 variables:
#>  $ vote  : Nmnl. item w/ 7 labels for 1,2,3,... + ms.v.  num  99 99 97 8 2 97 9 1 8 8 ...
#>  $ region: Nmnl. item w/ 5 labels for 1,2,3,... + ms.v.  num  2 2 99 99 1 99 1 2 2 2 ...
#>  $ income: Rto. item  num  432 3077 888 1417 2445 ...
summary(Data)
#>                    vote                     region        income       
#>  Conservatives       :38   England             :124   Min.   :  295.9  
#>  Labour              :47   Scotland            : 82   1st Qu.: 1202.0  
#>  Liberal Democrats   :38   Wales               : 40   Median : 2011.7  
#>  *Don't know         :42   *Not asked in survey: 54   Mean   : 2777.2  
#>  *Answer refused     :42                              3rd Qu.: 3541.5  
#>  *Not applicable     :43                              Max.   :23022.5  
#>  *Not asked in survey:50                                               

if (FALSE) {
# If we want to 'View' a data set we can use 'dsView'
dsView(Data)
# Works also, but changes the data set into a data frame first:
View(Data)
}

Data[[1]]
#> 
#> Item 'Vote intention' (measurement: nominal, type: double, length = 300) 
#> 
#>  [1:300] *Not asked in survey *Not asked in survey *Not applicable ...
Data[1,]
#> 
#> Data set with 1 observations and 3 variables
#> 
#>                   vote   region   income
#> 1 *Not asked in survey Scotland 431.9836
head(as.data.frame(Data))
#>     vote   region    income
#> 1   <NA> Scotland  431.9836
#> 2   <NA> Scotland 3076.6872
#> 3   <NA>     <NA>  887.9832
#> 4   <NA>     <NA> 1416.5029
#> 5 Labour  England 2445.3780
#> 6   <NA>     <NA> 2022.1012

EnglandData <- subset(Data,region == "England")
EnglandData
#> 
#> Data set with 124 observations and 3 variables
#> 
#>                    vote  region    income
#>  1               Labour England 2445.3780
#>  2      *Answer refused England  950.5021
#>  3      *Not applicable England 1202.7189
#>  4    Liberal Democrats England 3632.0289
#>  5 *Not asked in survey England 4548.7792
#>  6          *Don't know England 2097.1810
#>  7      *Not applicable England  655.0824
#>  8        Conservatives England 1067.6672
#>  9    Liberal Democrats England 4320.4525
#> 10               Labour England 1277.2426
#> 11 *Not asked in survey England 4481.2392
#> 12      *Not applicable England 1503.6296
#> 13          *Don't know England 1038.5655
#> 14        Conservatives England 2973.7063
#> 15               Labour England  726.7287
#> 16      *Answer refused England 2245.8013
#> 17      *Not applicable England 2984.2633
#> 18               Labour England 1282.2346
#> 19      *Not applicable England  684.1154
#> 20               Labour England 5464.8030
#> 21               Labour England  529.2084
#> 22               Labour England 3167.2084
#> 23        Conservatives England  992.9251
#> 24          *Don't know England 2443.0487
#> 25               Labour England 2381.2217
#> .. .................... ....... .........
#> (25 of 124 observations shown)

xtabs(~vote+region,data=Data)
#>                    region
#> vote                England Scotland Wales
#>   Conservatives          15       10     4
#>   Labour                 19       19     2
#>   Liberal Democrats      15        8     9
xtabs(~vote+region,data=within(Data, vote <- include.missings(vote)))
#>                       region
#> vote                   England Scotland Wales
#>   Conservatives             15       10     4
#>   Labour                    19       19     2
#>   Liberal Democrats         15        8     9
#>   *Don't know               12       16     5
#>   *Answer refused           19        9     6
#>   *Not applicable           18       11     6
#>   *Not asked in survey      26        9     8