Get Model Summaries for Use with "mtable"

A generic function and methods to collect coefficients and summary statistics from a model object. It is used in mtable

Usage

# S3 method for lm
getSummary(obj, alpha=.05,...)
  # S3 method for glm
getSummary(obj, alpha=.05,...)
  # S3 method for merMod
getSummary(obj, alpha=.05, ...)

# These are contributed by Christopher N. Lawrence
  # S3 method for clm
getSummary(obj, alpha=.05,...)
  # S3 method for polr
getSummary(obj, alpha=.05,...)
  # S3 method for simex
getSummary(obj, alpha=.05,...)

# These are contributed by Jason W. Morgan
  # S3 method for aftreg
getSummary(obj, alpha=.05,...)
  # S3 method for coxph
getSummary(obj, alpha=.05,...)
  # S3 method for phreg
getSummary(obj, alpha=.05,...)
  # S3 method for survreg
getSummary(obj, alpha=.05,...)
  # S3 method for weibreg
getSummary(obj, alpha=.05,...)

# These are contributed by Achim Zeileis
  # S3 method for ivreg
getSummary(obj, alpha=.05,...)
  # S3 method for tobit
getSummary(obj, alpha=.05,...)
  # S3 method for hurdle
getSummary(obj, alpha=.05,...)
  # S3 method for zeroinfl
getSummary(obj, alpha=.05,...)
  # S3 method for betareg
getSummary(obj, alpha=.05,...)
  # S3 method for multinom
getSummary(obj, alpha=.05,...)
  
# A variant that reports exponentiated coefficients.
# The default method calls 'getSummary()' internally and should
# be applicable to all classes for which 'getSummary()' methods exist.
getSummary_expcoef(obj, alpha=.05,...)
  # S3 method for default
getSummary_expcoef(obj, alpha=.05,...)

Arguments

obj

a model object, e.g. of class lm or glm

alpha

level of the confidence intervals; their coverage should be 1-alpha/2

...

further arguments; ignored.

Details

The generic function getSummary is called by mtable in order to obtain the coefficients and summaries of model objects. In order to adapt mtable to models of classes other than lm or glm one needs to define getSummary methods for these classes and to set a summary template via setSummaryTemplate

Value

Any method of getSummary must return a list with the following components:

coef

an array with coefficient estimates; the lowest dimension must have the following names and meanings:

est		the coefficient estimates,

se		the estimated standard errors,

stat		t- or Wald-z statistics,

p		significance levels of the statistics,

lwr		lower confidence limits,

upr		upper confidence limits.

The higher dimensions of the array correspond to the individual coefficients and, in multi-equation models, to the model equations.

sumstat

a vector containing the model summary statistics; the components may have arbitrary names.