Operate on grouped data in data frames and data sets

Group creates a grouped variant of an object of class "data.frame" or of class "data.set", for which methods for with and within are defined, so that these well-known functions can be applied "groupwise".

Usage

# Create an object of class "grouped.data" from a
# data frame or a data set.
Groups(data,by,...)
# S3 method for class 'data.frame'
Groups(data,by,...)
# S3 method for class 'data.set'
Groups(data,by,...)
# S3 method for class 'grouped.data'
Groups(data,by,...)

# Recombine grouped data into a data fame or a data set
recombine(x,...)
# S3 method for class 'grouped.data.frame'
recombine(x,...)
# S3 method for class 'grouped.data.set'
recombine(x,...)

# Recombine grouped data and coerce the result appropriately:
# S3 method for class 'grouped.data'
as.data.frame(x,...)
# S4 method for class 'grouped.data.frame'
as.data.set(x,row.names=NULL,...)
# S4 method for class 'grouped.data.set'
as.data.set(x,row.names=NULL,...)

# Methods of the generics "with" and "within" for grouped data
# S3 method for class 'grouped.data'
with(data,expr,...)
# S3 method for class 'grouped.data'
within(data,expr,recombine=FALSE,...)

# This is equivalent to with(Groups(data,by),expr,...)
withGroups(data,by,expr,...)
# This is equivalent to within(Groups(data,by),expr,recombine,...)
withinGroups(data,by,expr,recombine=TRUE,...)

Arguments

data

an object of the classes "data.frame", "data.set" if an argument to Groups, withGroups, withinGroups,

by

a formula with the factors the levels of which define the groups.

expr

an expression, or several expressions enclosed in curly braces.

recombine

a logical vector; should the resulting grouped data be recombined?

x

an object of class "grouped.data".

row.names

an optional character vector with row names.

...

other arguments, ignored.

Details

When applied to a data frame Groups returns an object with class attributes "grouped.data.frame", "grouped.data", and "data.frame", when applied do an object with class "data.set", it returns an object with class attributes "grouped.data.set", "grouped.data", and "data.set".

When applied to objects with class attributed "grouped.data", both the functions with() amd within() evaluate expr separately for each group defined by Groups. with() returns an array composed of the results of expr, while within() returns a modified copy of its data argument, which will be a "grouped.data" object ("grouped.data.frame" or "grouped.data.set"), unless the argument recombine=TRUE is set.

The expression expr may contain references to the variables n_, N_, and i_. n_ is equal to the size of the respective group (the number of rows belonging to it), while N_ is equal to the total number of observations in all groups. The variable i_ equals to the indices of the rows belonging to the respective group of observations.

Examples

some.data <- data.frame(x=rnorm(n=100))
some.data <- within(some.data,{
    f <- factor(rep(1:4,each=25),labels=letters[1:4])
    g <- factor(rep(1:5,each=4,5),labels=LETTERS[1:5])
    y <- x + rep(1:4,each=25) +  0.75*rep(1:5,each=4,5)
})

# For demonstration purposes, we create an
# 'empty' group:
some.data <- subset(some.data,
                       f!="a" | g!="C")

some.grouped.data <- Groups(some.data,
                           ~f+g)    

# Computing the means of y for each combination f and g
group.means <- with(some.grouped.data,
                    mean(y))
group.means
#>    g
#> f          A        B        C        D        E
#>   a 1.366103 2.313805       NA 4.015704 5.826801
#>   b 3.174157 3.566718 3.681667 4.955042 6.163240
#>   c 2.912945 4.288701 6.211069 5.887790 6.497435
#>   d 5.989396 6.232883 6.373659 8.116066 7.531311

# Obtaining a groupwise centered variant of y
some.grouped.data <- within(some.grouped.data,{
    y.cent <- y - mean(y)
},recombine=FALSE)

# The groupwise centered variable should have zero mean
# whithin each group
group.means <- with(some.grouped.data,
                    round(mean(y.cent),15))
group.means
#>    g
#> f   A B  C     D E
#>   a 0 0 NA 0e+00 0
#>   b 0 0  0 0e+00 0
#>   c 0 0  0 0e+00 0
#>   d 0 0  0 1e-15 0

# The following demonstrates the use of n_, N_, and i_
# An external copy of y
y1 <- some.data$y
group.means.n <- with(some.grouped.data,
                      c(mean(y),  # Group means for y
                        n_,       # Group sizes
                        sum(y)/n_,# Group means for y
                        n_/N_,    # Relative group sizes
                        sum(y1)/N_,# NOT the grand mean
                        sum(y1[i_])/n_)) # Group mean for y1
group.means.n
#> , , g = A
#> 
#>                 f
#>                           a          b          c          d
#>   mean(y)        1.36610256 3.17415667 2.91294505 5.98939584
#>   n_             8.00000000 4.00000000 4.00000000 4.00000000
#>   sum(y)/n_      1.36610256 3.17415667 2.91294505 5.98939584
#>   n_/N_          0.08333333 0.04166667 0.04166667 0.04166667
#>   sum(y1)/N_     4.94360924 4.94360924 4.94360924 4.94360924
#>   sum(y1[i_])/n_ 1.36610256 3.17415667 2.91294505 5.98939584
#> 
#> , , g = B
#> 
#>                 f
#>                           a          b          c          d
#>   mean(y)        2.31380497 3.56671812 4.28870074 6.23288284
#>   n_             5.00000000 7.00000000 4.00000000 4.00000000
#>   sum(y)/n_      2.31380497 3.56671812 4.28870074 6.23288284
#>   n_/N_          0.05208333 0.07291667 0.04166667 0.04166667
#>   sum(y1)/N_     4.94360924 4.94360924 4.94360924 4.94360924
#>   sum(y1[i_])/n_ 2.31380497 3.56671812 4.28870074 6.23288284
#> 
#> , , g = C
#> 
#>                 f
#>                   a        b        c          d
#>   mean(y)        NA 3.681667 6.211069 6.37365894
#>   n_             NA 6.000000 6.000000 4.00000000
#>   sum(y)/n_      NA 3.681667 6.211069 6.37365894
#>   n_/N_          NA 0.062500 0.062500 0.04166667
#>   sum(y1)/N_     NA 4.943609 4.943609 4.94360924
#>   sum(y1[i_])/n_ NA 3.681667 6.211069 6.37365894
#> 
#> , , g = D
#> 
#>                 f
#>                           a          b          c          d
#>   mean(y)        4.01570424 4.95504209 5.88779021 8.11606610
#>   n_             4.00000000 4.00000000 7.00000000 5.00000000
#>   sum(y)/n_      4.01570424 4.95504209 5.88779021 8.11606610
#>   n_/N_          0.04166667 0.04166667 0.07291667 0.05208333
#>   sum(y1)/N_     4.94360924 4.94360924 4.94360924 4.94360924
#>   sum(y1[i_])/n_ 4.01570424 4.95504209 5.88779021 8.11606610
#> 
#> , , g = E
#> 
#>                 f
#>                           a          b          c          d
#>   mean(y)        5.82680091 6.16323983 6.49743466 7.53131116
#>   n_             4.00000000 4.00000000 4.00000000 8.00000000
#>   sum(y)/n_      5.82680091 6.16323983 6.49743466 7.53131116
#>   n_/N_          0.04166667 0.04166667 0.04166667 0.08333333
#>   sum(y1)/N_     4.94360924 4.94360924 4.94360924 4.94360924
#>   sum(y1[i_])/n_ 5.82680091 6.16323983 6.49743466 7.53131116
#> 

# Names can be attached to the groupwise results
with(some.grouped.data,
     c(Centered=round(mean(y.cent),15),
       Uncentered=mean(y)))
#> , , g = A
#> 
#>             f
#>                     a        b        c        d
#>   Centered   0.000000 0.000000 0.000000 0.000000
#>   Uncentered 1.366103 3.174157 2.912945 5.989396
#> 
#> , , g = B
#> 
#>             f
#>                     a        b        c        d
#>   Centered   0.000000 0.000000 0.000000 0.000000
#>   Uncentered 2.313805 3.566718 4.288701 6.232883
#> 
#> , , g = C
#> 
#>             f
#>               a        b        c        d
#>   Centered   NA 0.000000 0.000000 0.000000
#>   Uncentered NA 3.681667 6.211069 6.373659
#> 
#> , , g = D
#> 
#>             f
#>                     a        b       c            d
#>   Centered   0.000000 0.000000 0.00000 1.000000e-15
#>   Uncentered 4.015704 4.955042 5.88779 8.116066e+00
#> 
#> , , g = E
#> 
#>             f
#>                     a       b        c        d
#>   Centered   0.000000 0.00000 0.000000 0.000000
#>   Uncentered 5.826801 6.16324 6.497435 7.531311
#> 

some.data.ungrouped <- recombine(some.grouped.data)
str(some.data.ungrouped)
#> 'data.frame':	96 obs. of  5 variables:
#>  $ x     : num  -1.756 -0.691 -0.559 -0.537 0.227 ...
#>  $ y     : num  -0.00628 1.05946 1.19146 1.21334 2.72713 ...
#>  $ g     : Factor w/ 5 levels "A","B","C","D",..: 1 1 1 1 2 2 2 2 4 4 ...
#>  $ f     : Factor w/ 4 levels "a","b","c","d": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ y.cent: num  -1.372 -0.307 -0.175 -0.153 0.413 ...

# It all works with "data.set" objects
some.dataset <- as.data.set(some.data)
some.grouped.dataset <- Groups(some.dataset,~f+g)

with(some.grouped.dataset,
     c(Mean=mean(y),
       Variance=var(y)))
#> , , g = A
#> 
#>           f
#>                    a        b         c        d
#>   Mean     1.3661026 3.174157 2.9129450 5.989396
#>   Variance 0.4896725 1.168783 0.8136362 1.578382
#> 
#> , , g = B
#> 
#>           f
#>                    a        b         c         d
#>   Mean     2.3138050 3.566718 4.2887007 6.2328828
#>   Variance 0.7791656 1.143553 0.3768418 0.5685829
#> 
#> , , g = C
#> 
#>           f
#>             a         b          c        d
#>   Mean     NA 3.6816674 6.21106857 6.373659
#>   Variance NA 0.4955471 0.04542602 1.360113
#> 
#> , , g = D
#> 
#>           f
#>                   a         b         c         d
#>   Mean     4.015704 4.9550421 5.8877902 8.1160661
#>   Variance 1.409835 0.5370272 0.5407879 0.5044257
#> 
#> , , g = E
#> 
#>           f
#>                    a        b        c        d
#>   Mean     5.8268009 6.163240 6.497435 7.531311
#>   Variance 0.6140468 1.378847 1.465161 1.323788
#> 

# The following two expressions are equivalent:
with(Groups(some.data,~f+g),mean(y))
#>    g
#> f          A        B        C        D        E
#>   a 1.366103 2.313805       NA 4.015704 5.826801
#>   b 3.174157 3.566718 3.681667 4.955042 6.163240
#>   c 2.912945 4.288701 6.211069 5.887790 6.497435
#>   d 5.989396 6.232883 6.373659 8.116066 7.531311
withGroups(some.data,~f+g,mean(y))
#>    g
#> f          A        B        C        D        E
#>   a 1.366103 2.313805       NA 4.015704 5.826801
#>   b 3.174157 3.566718 3.681667 4.955042 6.163240
#>   c 2.912945 4.288701 6.211069 5.887790 6.497435
#>   d 5.989396 6.232883 6.373659 8.116066 7.531311

# The following two expressions are equivalent:
some.data <- within(Groups(some.data,~f+g),{
    y.cent <- y - mean(y)
    y.cent.1 <- y - sum(y)/n_
})

some.data <- withinGroups(some.data,~f+g,{
    y.cent <- y - mean(y)
    y.cent.1 <- y - sum(y)/n_
})

# Both variants of groupwise centred varaibles should
# have zero groupwise means:
withGroups(some.data,~f+g,{
    c(round(mean(y.cent),15),
      round(mean(y.cent.1),15))
})
#> , , g = A
#> 
#>    f
#>     a b c d
#>   1 0 0 0 0
#>   2 0 0 0 0
#> 
#> , , g = B
#> 
#>    f
#>     a b c d
#>   1 0 0 0 0
#>   2 0 0 0 0
#> 
#> , , g = C
#> 
#>    f
#>      a b c d
#>   1 NA 0 0 0
#>   2 NA 0 0 0
#> 
#> , , g = D
#> 
#>    f
#>     a b      c      d
#>   1 0 0  0e+00  1e-15
#>   2 0 0 -1e-15 -1e-15
#> 
#> , , g = E
#> 
#>    f
#>     a b c d
#>   1 0 0 0 0
#>   2 0 0 0 0
#>