Data Set Objects

"data.set" objects are collections of "item" objects, with similar semantics as data frames. They are distinguished from data frames so that coercion by as.data.fame leads to a data frame that contains only vectors and factors. Nevertheless most methods for data frames are inherited by data sets, except for the method for the within generic function. For the within method for data sets, see the details section.

Thus data preparation using data sets retains all informations about item annotations, labels, missing values etc. While (mostly automatic) conversion of data sets into data frames makes the data amenable for the use of R's statistical functions.

dsView is a function that displays data sets in a similar manner as View displays data frames. (View works with data sets as well, but changes them first into data frames.)

Usage

data.set(...,row.names = NULL, check.rows = FALSE, check.names = TRUE,
    stringsAsFactors = FALSE, document = NULL)
as.data.set(x, row.names=NULL, ...)
# S4 method for list
as.data.set(x,row.names=NULL,...)
is.data.set(x)
# S3 method for data.set
as.data.frame(x, row.names = NULL, optional = FALSE, ...)
# S4 method for data.set
within(data, expr, ...)

dsView(x)

# S4 method for data.set
head(x,n=20,...)
# S4 method for data.set
tail(x,n=20,...)

Arguments

...

For the data.set function several vectors or items, for within further, ignored arguments.

row.names, check.rows, check.names, stringsAsFactors, optional

arguments as in data.frame or as.data.frame, respectively.

document

NULL or an optional character vector that contains documenation of the data.

x

for is.data.set(x), any object; for as.data.frame(x,...) and dsView(x) a "data.set" object.

data

a data set, that is, an object of class "data.set".

expr

an expression, or several expressions enclosed in curly braces.

n

integer; the number of rows to be shown by head or tail

Details

The as.data.frame method for data sets is just a copy of the method for list. Consequently, all items in the data set are coerced in accordance to their measurement setting, see as.vector,item-method and measurement.

The within method for data sets has the same effect as the within method for data frames, apart from two differences: all results of the computations are coerced into items if they have the appropriate length, otherwise, they are automatically dropped.

Currently only one method for the generic function as.data.set is defined: a method for "importer" objects.

Value

data.set and the within method for data sets returns a "data.set" object, is.data.set

returns a logical value, and as.data.frame returns a data frame.

Examples

Data <- data.set(
          vote = sample(c(1,2,3,8,9,97,99),size=300,replace=TRUE),
          region = sample(c(rep(1,3),rep(2,2),3,99),size=300,replace=TRUE),
          income = exp(rnorm(300,sd=.7))*2000
          )

Data <- within(Data,{
  description(vote) <- "Vote intention"
  description(region) <- "Region of residence"
  description(income) <- "Household income"
  wording(vote) <- "If a general election would take place next tuesday,
                    the candidate of which party would you vote for?"
  wording(income) <- "All things taken into account, how much do all
                    household members earn in sum?"
  foreach(x=c(vote,region),{
    measurement(x) <- "nominal"
    })
  measurement(income) <- "ratio"
  labels(vote) <- c(
                    Conservatives         =  1,
                    Labour                =  2,
                    "Liberal Democrats"   =  3,
                    "Don't know"          =  8,
                    "Answer refused"      =  9,
                    "Not applicable"      = 97,
                    "Not asked in survey" = 99)
  labels(region) <- c(
                    England               =  1,
                    Scotland              =  2,
                    Wales                 =  3,
                    "Not applicable"      = 97,
                    "Not asked in survey" = 99)
  foreach(x=c(vote,region,income),{
    annotation(x)["Remark"] <- "This is not a real survey item, of course ..."
    })
  missing.values(vote) <- c(8,9,97,99)
  missing.values(region) <- c(97,99)

  # These to variables do not appear in the
  # the resulting data set, since they have the wrong length.
  junk1 <- 1:5
  junk2 <- matrix(5,4,4)
  
})
#> Warning: Variables 'junk1','junk2' have wrong length, removing them.
# Since data sets may be huge, only a
# part of them are 'show'n
Data
#> 
#> Data set with 300 observations and 3 variables
#> 
#>                    vote               region     income
#>  1 *Not asked in survey *Not asked in survey  1715.6085
#>  2      *Not applicable *Not asked in survey  2252.1969
#>  3               Labour *Not asked in survey  3877.5634
#>  4        Conservatives                Wales   957.1327
#>  5      *Not applicable              England  5828.3387
#>  6    Liberal Democrats *Not asked in survey 15438.2770
#>  7        Conservatives              England  1736.9046
#>  8        Conservatives             Scotland  1343.7010
#>  9               Labour              England  1881.3877
#> 10          *Don't know                Wales  3372.9566
#> 11      *Not applicable              England  2709.0086
#> 12          *Don't know             Scotland  1550.5605
#> 13    Liberal Democrats              England  1782.3988
#> 14          *Don't know                Wales  1088.8877
#> 15 *Not asked in survey *Not asked in survey  1568.4711
#> 16               Labour              England  1673.8212
#> 17          *Don't know              England   704.1890
#> 18        Conservatives *Not asked in survey  1023.8533
#> 19        Conservatives              England  5322.1834
#> 20      *Not applicable              England  1927.4379
#> 21      *Answer refused              England  5909.2853
#> 22      *Not applicable             Scotland  3377.8488
#> 23      *Answer refused              England  4466.9434
#> 24          *Don't know              England  4842.7757
#> 25 *Not asked in survey             Scotland  1271.7488
#> .. .................... .................... ..........
#> (25 of 300 observations shown)

if (FALSE) {

# If we insist on seeing all, we can use 'print' instead
print(Data)
}

str(Data)
#> Data set with 300 obs. of 3 variables:
#>  $ vote  : Nmnl. item w/ 7 labels for 1,2,3,... + ms.v.  num  99 97 2 1 97 3 1 1 2 8 ...
#>  $ region: Nmnl. item w/ 5 labels for 1,2,3,... + ms.v.  num  99 99 99 3 1 99 1 2 1 3 ...
#>  $ income: Rto. item  num  1716 2252 3878 957 5828 ...
summary(Data)
#>                    vote                     region        income       
#>  Conservatives       :44   England             :122   Min.   :  241.1  
#>  Labour              :56   Scotland            : 75   1st Qu.: 1309.3  
#>  Liberal Democrats   :45   Wales               : 49   Median : 2016.5  
#>  *Don't know         :31   *Not asked in survey: 54   Mean   : 2693.4  
#>  *Answer refused     :48                              3rd Qu.: 3293.5  
#>  *Not applicable     :41                              Max.   :15438.3  
#>  *Not asked in survey:35                                               

if (FALSE) {
# If we want to 'View' a data set we can use 'dsView'
dsView(Data)
# Works also, but changes the data set into a data frame first:
View(Data)
}

Data[[1]]
#> 
#> Item 'Vote intention' (measurement: nominal, type: double, length = 300) 
#> 
#>  [1:300] *Not asked in survey *Not applicable Labour Conservatives ...
Data[1,]
#> 
#> Data set with 1 observations and 3 variables
#> 
#>                   vote               region   income
#> 1 *Not asked in survey *Not asked in survey 1715.609
head(as.data.frame(Data))
#>                vote  region     income
#> 1              <NA>    <NA>  1715.6085
#> 2              <NA>    <NA>  2252.1969
#> 3            Labour    <NA>  3877.5634
#> 4     Conservatives   Wales   957.1327
#> 5              <NA> England  5828.3387
#> 6 Liberal Democrats    <NA> 15438.2770

EnglandData <- subset(Data,region == "England")
EnglandData
#> 
#> Data set with 122 observations and 3 variables
#> 
#>                 vote  region    income
#>  1   *Not applicable England 5828.3387
#>  2     Conservatives England 1736.9046
#>  3            Labour England 1881.3877
#>  4   *Not applicable England 2709.0086
#>  5 Liberal Democrats England 1782.3988
#>  6            Labour England 1673.8212
#>  7       *Don't know England  704.1890
#>  8     Conservatives England 5322.1834
#>  9   *Not applicable England 1927.4379
#> 10   *Answer refused England 5909.2853
#> 11   *Answer refused England 4466.9434
#> 12       *Don't know England 4842.7757
#> 13   *Answer refused England 2103.0118
#> 14     Conservatives England  684.7497
#> 15     Conservatives England 2865.5092
#> 16 Liberal Democrats England 1621.3743
#> 17            Labour England 1387.0668
#> 18 Liberal Democrats England 3318.2035
#> 19   *Answer refused England 1957.5920
#> 20   *Answer refused England 2743.2113
#> 21            Labour England 6947.0691
#> 22     Conservatives England 9186.9795
#> 23 Liberal Democrats England 3130.7817
#> 24   *Answer refused England 1131.3532
#> 25            Labour England 3170.6028
#> .. ................. ....... .........
#> (25 of 122 observations shown)

xtabs(~vote+region,data=Data)
#>                    region
#> vote                England Scotland Wales
#>   Conservatives          17       16     4
#>   Labour                 30       11     7
#>   Liberal Democrats      18       13     9
xtabs(~vote+region,data=within(Data, vote <- include.missings(vote)))
#>                       region
#> vote                   England Scotland Wales
#>   Conservatives             17       16     4
#>   Labour                    30       11     7
#>   Liberal Democrats         18       13     9
#>   *Don't know               11        4     9
#>   *Answer refused           23        8     8
#>   *Not applicable           14       13     8
#>   *Not asked in survey       9       10     4