Data Set Objects

"data.set" objects are collections of "item" objects, with similar semantics as data frames. They are distinguished from data frames so that coercion by as.data.fame leads to a data frame that contains only vectors and factors. Nevertheless most methods for data frames are inherited by data sets, except for the method for the within generic function. For the within method for data sets, see the details section.

Thus data preparation using data sets retains all informations about item annotations, labels, missing values etc. While (mostly automatic) conversion of data sets into data frames makes the data amenable for the use of R's statistical functions.

dsView is a function that displays data sets in a similar manner as View displays data frames. (View works with data sets as well, but changes them first into data frames.)

Usage

data.set(...,row.names = NULL, check.rows = FALSE, check.names = TRUE,
    stringsAsFactors = FALSE, document = NULL)
as.data.set(x, row.names=NULL, ...)
# S4 method for list
as.data.set(x,row.names=NULL,...)
is.data.set(x)
# S3 method for data.set
as.data.frame(x, row.names = NULL, optional = FALSE, ...)
# S4 method for data.set
within(data, expr, ...)

dsView(x)

# S4 method for data.set
head(x,n=20,...)
# S4 method for data.set
tail(x,n=20,...)

Arguments

...

For the data.set function several vectors or items, for within further, ignored arguments.

row.names, check.rows, check.names, stringsAsFactors, optional

arguments as in data.frame or as.data.frame, respectively.

document

NULL or an optional character vector that contains documenation of the data.

x

for is.data.set(x), any object; for as.data.frame(x,...) and dsView(x) a "data.set" object.

data

a data set, that is, an object of class "data.set".

expr

an expression, or several expressions enclosed in curly braces.

n

integer; the number of rows to be shown by head or tail

Details

The as.data.frame method for data sets is just a copy of the method for list. Consequently, all items in the data set are coerced in accordance to their measurement setting, see as.vector,item-method and measurement.

The within method for data sets has the same effect as the within method for data frames, apart from two differences: all results of the computations are coerced into items if they have the appropriate length, otherwise, they are automatically dropped.

Currently only one method for the generic function as.data.set is defined: a method for "importer" objects.

Value

data.set and the within method for data sets returns a "data.set" object, is.data.set

returns a logical value, and as.data.frame returns a data frame.

Examples

Data <- data.set(
          vote = sample(c(1,2,3,8,9,97,99),size=300,replace=TRUE),
          region = sample(c(rep(1,3),rep(2,2),3,99),size=300,replace=TRUE),
          income = exp(rnorm(300,sd=.7))*2000
          )

Data <- within(Data,{
  description(vote) <- "Vote intention"
  description(region) <- "Region of residence"
  description(income) <- "Household income"
  wording(vote) <- "If a general election would take place next tuesday,
                    the candidate of which party would you vote for?"
  wording(income) <- "All things taken into account, how much do all
                    household members earn in sum?"
  foreach(x=c(vote,region),{
    measurement(x) <- "nominal"
    })
  measurement(income) <- "ratio"
  labels(vote) <- c(
                    Conservatives         =  1,
                    Labour                =  2,
                    "Liberal Democrats"   =  3,
                    "Don't know"          =  8,
                    "Answer refused"      =  9,
                    "Not applicable"      = 97,
                    "Not asked in survey" = 99)
  labels(region) <- c(
                    England               =  1,
                    Scotland              =  2,
                    Wales                 =  3,
                    "Not applicable"      = 97,
                    "Not asked in survey" = 99)
  foreach(x=c(vote,region,income),{
    annotation(x)["Remark"] <- "This is not a real survey item, of course ..."
    })
  missing.values(vote) <- c(8,9,97,99)
  missing.values(region) <- c(97,99)

  # These to variables do not appear in the
  # the resulting data set, since they have the wrong length.
  junk1 <- 1:5
  junk2 <- matrix(5,4,4)
  
})
#> Warning: Variables 'junk1','junk2' have wrong length, removing them.
# Since data sets may be huge, only a
# part of them are 'show'n
Data
#> 
#> Data set with 300 observations and 3 variables
#> 
#>                    vote               region     income
#>  1 *Not asked in survey              England 12076.0588
#>  2        Conservatives *Not asked in survey  3270.8117
#>  3               Labour                Wales  2765.8968
#>  4      *Not applicable              England   668.8680
#>  5      *Not applicable              England  2653.1078
#>  6          *Don't know              England   279.2010
#>  7      *Not applicable                Wales  1328.9843
#>  8        Conservatives *Not asked in survey  2073.2958
#>  9      *Answer refused              England  1658.4523
#> 10      *Answer refused             Scotland   632.3376
#> 11          *Don't know             Scotland   985.2760
#> 12      *Answer refused              England  4128.1992
#> 13          *Don't know              England  8083.8686
#> 14      *Answer refused             Scotland  4008.2079
#> 15               Labour              England  2273.4031
#> 16 *Not asked in survey              England  2637.8494
#> 17      *Answer refused             Scotland  1357.0114
#> 18      *Not applicable                Wales  1838.1643
#> 19    Liberal Democrats              England  1731.1915
#> 20        Conservatives *Not asked in survey   977.2703
#> 21    Liberal Democrats             Scotland  4828.5276
#> 22        Conservatives              England  2452.8176
#> 23          *Don't know              England  1558.5283
#> 24      *Not applicable                Wales  4791.8016
#> 25               Labour                Wales  2019.0097
#> .. .................... .................... ..........
#> (25 of 300 observations shown)

if (FALSE) {

# If we insist on seeing all, we can use 'print' instead
print(Data)
}

str(Data)
#> Data set with 300 obs. of 3 variables:
#>  $ vote  : Nmnl. item w/ 7 labels for 1,2,3,... + ms.v.  num  99 1 2 97 97 8 97 1 9 9 ...
#>  $ region: Nmnl. item w/ 5 labels for 1,2,3,... + ms.v.  num  1 99 3 1 1 1 3 99 1 2 ...
#>  $ income: Rto. item  num  12076 3271 2766 669 2653 ...
summary(Data)
#>                    vote                     region        income       
#>  Conservatives       :41   England             :126   Min.   :  279.2  
#>  Labour              :40   Scotland            : 92   1st Qu.: 1222.3  
#>  Liberal Democrats   :40   Wales               : 46   Median : 1945.4  
#>  *Don't know         :45   *Not asked in survey: 36   Mean   : 2451.7  
#>  *Answer refused     :56                              3rd Qu.: 2801.8  
#>  *Not applicable     :45                              Max.   :12076.1  
#>  *Not asked in survey:33                                               

if (FALSE) {
# If we want to 'View' a data set we can use 'dsView'
dsView(Data)
# Works also, but changes the data set into a data frame first:
View(Data)
}

Data[[1]]
#> 
#> Item 'Vote intention' (measurement: nominal, type: double, length = 300) 
#> 
#>  [1:300] *Not asked in survey Conservatives Labour *Not applicable ...
Data[1,]
#> 
#> Data set with 1 observations and 3 variables
#> 
#>                   vote  region   income
#> 1 *Not asked in survey England 12076.06
head(as.data.frame(Data))
#>            vote  region    income
#> 1          <NA> England 12076.059
#> 2 Conservatives    <NA>  3270.812
#> 3        Labour   Wales  2765.897
#> 4          <NA> England   668.868
#> 5          <NA> England  2653.108
#> 6          <NA> England   279.201

EnglandData <- subset(Data,region == "England")
EnglandData
#> 
#> Data set with 126 observations and 3 variables
#> 
#>                    vote  region     income
#>  1 *Not asked in survey England 12076.0588
#>  2      *Not applicable England   668.8680
#>  3      *Not applicable England  2653.1078
#>  4          *Don't know England   279.2010
#>  5      *Answer refused England  1658.4523
#>  6      *Answer refused England  4128.1992
#>  7          *Don't know England  8083.8686
#>  8               Labour England  2273.4031
#>  9 *Not asked in survey England  2637.8494
#> 10    Liberal Democrats England  1731.1915
#> 11        Conservatives England  2452.8176
#> 12          *Don't know England  1558.5283
#> 13      *Not applicable England  2879.3263
#> 14          *Don't know England  1420.6832
#> 15      *Answer refused England  2063.6379
#> 16          *Don't know England  4067.1603
#> 17      *Answer refused England  2170.0382
#> 18               Labour England  2566.1221
#> 19      *Not applicable England   855.2873
#> 20      *Answer refused England   723.8749
#> 21          *Don't know England  2811.9149
#> 22      *Not applicable England  1971.7155
#> 23      *Answer refused England  1888.8399
#> 24    Liberal Democrats England  5507.4907
#> 25        Conservatives England  1810.2119
#> .. .................... ....... ..........
#> (25 of 126 observations shown)

xtabs(~vote+region,data=Data)
#>                    region
#> vote                England Scotland Wales
#>   Conservatives          18       13     5
#>   Labour                 13       11    12
#>   Liberal Democrats      17       12     3
xtabs(~vote+region,data=within(Data, vote <- include.missings(vote)))
#>                       region
#> vote                   England Scotland Wales
#>   Conservatives             18       13     5
#>   Labour                    13       11    12
#>   Liberal Democrats         17       12     3
#>   *Don't know               23       14     5
#>   *Answer refused           19       21     9
#>   *Not applicable           19       10     9
#>   *Not asked in survey      17       11     3