Data Set Objects

"data.set" objects are collections of "item" objects, with similar semantics as data frames. They are distinguished from data frames so that coercion by as.data.fame leads to a data frame that contains only vectors and factors. Nevertheless most methods for data frames are inherited by data sets, except for the method for the within generic function. For the within method for data sets, see the details section.

Thus data preparation using data sets retains all informations about item annotations, labels, missing values etc. While (mostly automatic) conversion of data sets into data frames makes the data amenable for the use of R's statistical functions.

dsView is a function that displays data sets in a similar manner as View displays data frames. (View works with data sets as well, but changes them first into data frames.)

Usage

data.set(...,row.names = NULL, check.rows = FALSE, check.names = TRUE,
    stringsAsFactors = FALSE, document = NULL)
as.data.set(x, row.names=NULL, ...)
# S4 method for list
as.data.set(x,row.names=NULL,...)
is.data.set(x)
# S3 method for data.set
as.data.frame(x, row.names = NULL, optional = FALSE, ...)
# S4 method for data.set
within(data, expr, ...)

dsView(x)

# S4 method for data.set
head(x,n=20,...)
# S4 method for data.set
tail(x,n=20,...)

Arguments

...

For the data.set function several vectors or items, for within further, ignored arguments.

row.names, check.rows, check.names, stringsAsFactors, optional

arguments as in data.frame or as.data.frame, respectively.

document

NULL or an optional character vector that contains documenation of the data.

x

for is.data.set(x), any object; for as.data.frame(x,...) and dsView(x) a "data.set" object.

data

a data set, that is, an object of class "data.set".

expr

an expression, or several expressions enclosed in curly braces.

n

integer; the number of rows to be shown by head or tail

Details

The as.data.frame method for data sets is just a copy of the method for list. Consequently, all items in the data set are coerced in accordance to their measurement setting, see as.vector,item-method and measurement.

The within method for data sets has the same effect as the within method for data frames, apart from two differences: all results of the computations are coerced into items if they have the appropriate length, otherwise, they are automatically dropped.

Currently only one method for the generic function as.data.set is defined: a method for "importer" objects.

Value

data.set and the within method for data sets returns a "data.set" object, is.data.set

returns a logical value, and as.data.frame returns a data frame.

Examples

Data <- data.set(
          vote = sample(c(1,2,3,8,9,97,99),size=300,replace=TRUE),
          region = sample(c(rep(1,3),rep(2,2),3,99),size=300,replace=TRUE),
          income = exp(rnorm(300,sd=.7))*2000
          )

Data <- within(Data,{
  description(vote) <- "Vote intention"
  description(region) <- "Region of residence"
  description(income) <- "Household income"
  wording(vote) <- "If a general election would take place next tuesday,
                    the candidate of which party would you vote for?"
  wording(income) <- "All things taken into account, how much do all
                    household members earn in sum?"
  foreach(x=c(vote,region),{
    measurement(x) <- "nominal"
    })
  measurement(income) <- "ratio"
  labels(vote) <- c(
                    Conservatives         =  1,
                    Labour                =  2,
                    "Liberal Democrats"   =  3,
                    "Don't know"          =  8,
                    "Answer refused"      =  9,
                    "Not applicable"      = 97,
                    "Not asked in survey" = 99)
  labels(region) <- c(
                    England               =  1,
                    Scotland              =  2,
                    Wales                 =  3,
                    "Not applicable"      = 97,
                    "Not asked in survey" = 99)
  foreach(x=c(vote,region,income),{
    annotation(x)["Remark"] <- "This is not a real survey item, of course ..."
    })
  missing.values(vote) <- c(8,9,97,99)
  missing.values(region) <- c(97,99)

  # These to variables do not appear in the
  # the resulting data set, since they have the wrong length.
  junk1 <- 1:5
  junk2 <- matrix(5,4,4)
  
})
#> Warning: Variables 'junk1','junk2' have wrong length, removing them.
# Since data sets may be huge, only a
# part of them are 'show'n
Data
#> 
#> Data set with 300 observations and 3 variables
#> 
#>                    vote               region     income
#>  1      *Not applicable              England  1379.5494
#>  2        Conservatives             Scotland  1460.5068
#>  3        Conservatives *Not asked in survey  2918.9946
#>  4          *Don't know              England  2397.0241
#>  5        Conservatives              England  2312.5063
#>  6      *Not applicable             Scotland  4299.0954
#>  7          *Don't know             Scotland  2233.1853
#>  8      *Not applicable              England  4487.4607
#>  9      *Answer refused                Wales  3750.4843
#> 10          *Don't know             Scotland  1858.8441
#> 11      *Not applicable              England  1107.3698
#> 12          *Don't know             Scotland   450.4971
#> 13               Labour             Scotland  2211.9403
#> 14      *Not applicable              England  1713.2538
#> 15          *Don't know             Scotland  5532.0869
#> 16      *Answer refused              England 18161.3228
#> 17      *Not applicable              England  4446.4739
#> 18    Liberal Democrats             Scotland   889.7442
#> 19          *Don't know              England  2986.9189
#> 20          *Don't know                Wales  7305.6285
#> 21 *Not asked in survey              England   338.4287
#> 22    Liberal Democrats              England  1668.9962
#> 23      *Answer refused             Scotland  1086.7915
#> 24      *Answer refused             Scotland   791.2063
#> 25      *Answer refused             Scotland  1569.8879
#> .. .................... .................... ..........
#> (25 of 300 observations shown)

if (FALSE) {

# If we insist on seeing all, we can use 'print' instead
print(Data)
}

str(Data)
#> Data set with 300 obs. of 3 variables:
#>  $ vote  : Nmnl. item w/ 7 labels for 1,2,3,... + ms.v.  num  97 1 1 8 1 97 8 97 9 8 ...
#>  $ region: Nmnl. item w/ 5 labels for 1,2,3,... + ms.v.  num  1 2 99 1 1 2 2 1 3 2 ...
#>  $ income: Rto. item  num  1380 1461 2919 2397 2313 ...
summary(Data)
#>                    vote                     region        income       
#>  Conservatives       :30   England             :144   Min.   :  147.8  
#>  Labour              :37   Scotland            : 75   1st Qu.: 1250.6  
#>  Liberal Democrats   :39   Wales               : 38   Median : 2148.2  
#>  *Don't know         :45   *Not asked in survey: 43   Mean   : 2638.6  
#>  *Answer refused     :48                              3rd Qu.: 3296.6  
#>  *Not applicable     :52                              Max.   :22312.1  
#>  *Not asked in survey:49                                               

if (FALSE) {
# If we want to 'View' a data set we can use 'dsView'
dsView(Data)
# Works also, but changes the data set into a data frame first:
View(Data)
}

Data[[1]]
#> 
#> Item 'Vote intention' (measurement: nominal, type: double, length = 300) 
#> 
#>  [1:300] *Not applicable Conservatives Conservatives *Don't know ...
Data[1,]
#> 
#> Data set with 1 observations and 3 variables
#> 
#>              vote  region   income
#> 1 *Not applicable England 1379.549
head(as.data.frame(Data))
#>            vote   region   income
#> 1          <NA>  England 1379.549
#> 2 Conservatives Scotland 1460.507
#> 3 Conservatives     <NA> 2918.995
#> 4          <NA>  England 2397.024
#> 5 Conservatives  England 2312.506
#> 6          <NA> Scotland 4299.095

EnglandData <- subset(Data,region == "England")
EnglandData
#> 
#> Data set with 144 observations and 3 variables
#> 
#>                    vote  region     income
#>  1      *Not applicable England  1379.5494
#>  2          *Don't know England  2397.0241
#>  3        Conservatives England  2312.5063
#>  4      *Not applicable England  4487.4607
#>  5      *Not applicable England  1107.3698
#>  6      *Not applicable England  1713.2538
#>  7      *Answer refused England 18161.3228
#>  8      *Not applicable England  4446.4739
#>  9          *Don't know England  2986.9189
#> 10 *Not asked in survey England   338.4287
#> 11    Liberal Democrats England  1668.9962
#> 12    Liberal Democrats England  3880.1077
#> 13      *Not applicable England  3192.0153
#> 14        Conservatives England   801.5920
#> 15      *Answer refused England  1619.8943
#> 16      *Not applicable England  1910.5419
#> 17               Labour England  2120.5616
#> 18      *Not applicable England 10759.2713
#> 19      *Not applicable England  4370.6836
#> 20      *Not applicable England  3085.8287
#> 21      *Not applicable England  5904.8352
#> 22          *Don't know England  2263.4971
#> 23      *Answer refused England  4231.4213
#> 24      *Answer refused England  2231.4009
#> 25 *Not asked in survey England  4329.2830
#> .. .................... ....... ..........
#> (25 of 144 observations shown)

xtabs(~vote+region,data=Data)
#>                    region
#> vote                England Scotland Wales
#>   Conservatives          11        9     3
#>   Labour                 15       10     6
#>   Liberal Democrats      19       10     5
xtabs(~vote+region,data=within(Data, vote <- include.missings(vote)))
#>                       region
#> vote                   England Scotland Wales
#>   Conservatives             11        9     3
#>   Labour                    15       10     6
#>   Liberal Democrats         19       10     5
#>   *Don't know               19        9    10
#>   *Answer refused           26       12     5
#>   *Not applicable           29       14     6
#>   *Not asked in survey      25       11     3