Data Set Objects

"data.set" objects are collections of "item" objects, with similar semantics as data frames. They are distinguished from data frames so that coercion by as.data.fame leads to a data frame that contains only vectors and factors. Nevertheless most methods for data frames are inherited by data sets, except for the method for the within generic function. For the within method for data sets, see the details section.

Thus data preparation using data sets retains all informations about item annotations, labels, missing values etc. While (mostly automatic) conversion of data sets into data frames makes the data amenable for the use of R's statistical functions.

dsView is a function that displays data sets in a similar manner as View displays data frames. (View works with data sets as well, but changes them first into data frames.)

Usage

data.set(...,row.names = NULL, check.rows = FALSE, check.names = TRUE,
    stringsAsFactors = FALSE, document = NULL)
as.data.set(x, row.names=NULL, ...)
# S4 method for list
as.data.set(x,row.names=NULL,...)
is.data.set(x)
# S3 method for data.set
as.data.frame(x, row.names = NULL, optional = FALSE, ...)
# S4 method for data.set
within(data, expr, ...)

dsView(x)

# S4 method for data.set
head(x,n=20,...)
# S4 method for data.set
tail(x,n=20,...)

Arguments

...

For the data.set function several vectors or items, for within further, ignored arguments.

row.names, check.rows, check.names, stringsAsFactors, optional

arguments as in data.frame or as.data.frame, respectively.

document

NULL or an optional character vector that contains documenation of the data.

x

for is.data.set(x), any object; for as.data.frame(x,...) and dsView(x) a "data.set" object.

data

a data set, that is, an object of class "data.set".

expr

an expression, or several expressions enclosed in curly braces.

n

integer; the number of rows to be shown by head or tail

Details

The as.data.frame method for data sets is just a copy of the method for list. Consequently, all items in the data set are coerced in accordance to their measurement setting, see as.vector,item-method and measurement.

The within method for data sets has the same effect as the within method for data frames, apart from two differences: all results of the computations are coerced into items if they have the appropriate length, otherwise, they are automatically dropped.

Currently only one method for the generic function as.data.set is defined: a method for "importer" objects.

Value

data.set and the within method for data sets returns a "data.set" object, is.data.set

returns a logical value, and as.data.frame returns a data frame.

Examples

Data <- data.set(
          vote = sample(c(1,2,3,8,9,97,99),size=300,replace=TRUE),
          region = sample(c(rep(1,3),rep(2,2),3,99),size=300,replace=TRUE),
          income = exp(rnorm(300,sd=.7))*2000
          )

Data <- within(Data,{
  description(vote) <- "Vote intention"
  description(region) <- "Region of residence"
  description(income) <- "Household income"
  wording(vote) <- "If a general election would take place next tuesday,
                    the candidate of which party would you vote for?"
  wording(income) <- "All things taken into account, how much do all
                    household members earn in sum?"
  foreach(x=c(vote,region),{
    measurement(x) <- "nominal"
    })
  measurement(income) <- "ratio"
  labels(vote) <- c(
                    Conservatives         =  1,
                    Labour                =  2,
                    "Liberal Democrats"   =  3,
                    "Don't know"          =  8,
                    "Answer refused"      =  9,
                    "Not applicable"      = 97,
                    "Not asked in survey" = 99)
  labels(region) <- c(
                    England               =  1,
                    Scotland              =  2,
                    Wales                 =  3,
                    "Not applicable"      = 97,
                    "Not asked in survey" = 99)
  foreach(x=c(vote,region,income),{
    annotation(x)["Remark"] <- "This is not a real survey item, of course ..."
    })
  missing.values(vote) <- c(8,9,97,99)
  missing.values(region) <- c(97,99)

  # These to variables do not appear in the
  # the resulting data set, since they have the wrong length.
  junk1 <- 1:5
  junk2 <- matrix(5,4,4)
  
})
#> Warning: Variables 'junk1','junk2' have wrong length, removing them.
# Since data sets may be huge, only a
# part of them are 'show'n
Data
#> 
#> Data set with 300 observations and 3 variables
#> 
#>                    vote               region     income
#>  1        Conservatives              England  1753.7972
#>  2        Conservatives             Scotland   337.5398
#>  3    Liberal Democrats             Scotland   492.8090
#>  4      *Answer refused             Scotland  1602.1649
#>  5    Liberal Democrats             Scotland  2382.3407
#>  6               Labour                Wales   352.4050
#>  7      *Answer refused              England  1284.0083
#>  8    Liberal Democrats              England  3579.2890
#>  9      *Answer refused              England  1164.3002
#> 10        Conservatives              England   824.4249
#> 11 *Not asked in survey              England  3201.0240
#> 12    Liberal Democrats             Scotland  2351.5873
#> 13        Conservatives *Not asked in survey  1724.9266
#> 14    Liberal Democrats             Scotland  2547.6688
#> 15               Labour *Not asked in survey  6562.0628
#> 16 *Not asked in survey              England  9344.5974
#> 17        Conservatives                Wales  3736.2437
#> 18               Labour *Not asked in survey  3641.4993
#> 19               Labour             Scotland  1275.4481
#> 20 *Not asked in survey *Not asked in survey  1603.0470
#> 21      *Answer refused              England  1725.6809
#> 22               Labour             Scotland  1121.7540
#> 23      *Not applicable              England 21888.1023
#> 24      *Answer refused             Scotland  1035.3694
#> 25    Liberal Democrats                Wales  1269.4865
#> .. .................... .................... ..........
#> (25 of 300 observations shown)

if (FALSE) {

# If we insist on seeing all, we can use 'print' instead
print(Data)
}

str(Data)
#> Data set with 300 obs. of 3 variables:
#>  $ vote  : Nmnl. item w/ 7 labels for 1,2,3,... + ms.v.  num  1 1 3 9 3 2 9 3 9 1 ...
#>  $ region: Nmnl. item w/ 5 labels for 1,2,3,... + ms.v.  num  1 2 2 2 2 3 1 1 1 1 ...
#>  $ income: Rto. item  num  1754 338 493 1602 2382 ...
summary(Data)
#>                    vote                     region        income       
#>  Conservatives       :48   England             :127   Min.   :  175.1  
#>  Labour              :48   Scotland            : 83   1st Qu.: 1328.3  
#>  Liberal Democrats   :39   Wales               : 46   Median : 2059.1  
#>  *Don't know         :40   *Not asked in survey: 44   Mean   : 2808.8  
#>  *Answer refused     :51                              3rd Qu.: 3311.5  
#>  *Not applicable     :37                              Max.   :21888.1  
#>  *Not asked in survey:37                                               

if (FALSE) {
# If we want to 'View' a data set we can use 'dsView'
dsView(Data)
# Works also, but changes the data set into a data frame first:
View(Data)
}

Data[[1]]
#> 
#> Item 'Vote intention' (measurement: nominal, type: double, length = 300) 
#> 
#>  [1:300] Conservatives Conservatives Liberal Democrats *Answer refused ...
Data[1,]
#> 
#> Data set with 1 observations and 3 variables
#> 
#>            vote  region   income
#> 1 Conservatives England 1753.797
head(as.data.frame(Data))
#>                vote   region    income
#> 1     Conservatives  England 1753.7972
#> 2     Conservatives Scotland  337.5398
#> 3 Liberal Democrats Scotland  492.8090
#> 4              <NA> Scotland 1602.1649
#> 5 Liberal Democrats Scotland 2382.3407
#> 6            Labour    Wales  352.4050

EnglandData <- subset(Data,region == "England")
EnglandData
#> 
#> Data set with 127 observations and 3 variables
#> 
#>                    vote  region     income
#>  1        Conservatives England  1753.7972
#>  2      *Answer refused England  1284.0083
#>  3    Liberal Democrats England  3579.2890
#>  4      *Answer refused England  1164.3002
#>  5        Conservatives England   824.4249
#>  6 *Not asked in survey England  3201.0240
#>  7 *Not asked in survey England  9344.5974
#>  8      *Answer refused England  1725.6809
#>  9      *Not applicable England 21888.1023
#> 10          *Don't know England  2037.9098
#> 11      *Answer refused England  3471.0778
#> 12      *Answer refused England  2453.4826
#> 13        Conservatives England  2172.8334
#> 14          *Don't know England 12880.8168
#> 15    Liberal Democrats England  2634.6286
#> 16        Conservatives England  2067.7439
#> 17      *Answer refused England  2611.6353
#> 18               Labour England  5496.1471
#> 19          *Don't know England  3284.3544
#> 20          *Don't know England  3184.1759
#> 21        Conservatives England  1115.1103
#> 22        Conservatives England  1743.2755
#> 23        Conservatives England  4144.6089
#> 24               Labour England  1418.2612
#> 25               Labour England  2566.3583
#> .. .................... ....... ..........
#> (25 of 127 observations shown)

xtabs(~vote+region,data=Data)
#>                    region
#> vote                England Scotland Wales
#>   Conservatives          21       15     8
#>   Labour                 16       16     9
#>   Liberal Democrats      15       18     2
xtabs(~vote+region,data=within(Data, vote <- include.missings(vote)))
#>                       region
#> vote                   England Scotland Wales
#>   Conservatives             21       15     8
#>   Labour                    16       16     9
#>   Liberal Democrats         15       18     2
#>   *Don't know               17        8     9
#>   *Answer refused           25       10     6
#>   *Not applicable           15        9     3
#>   *Not asked in survey      18        7     9