Reorder an Array or Matrix

reorder.array reorders an array along a specified dimension according given names, indices or results of a function applied.

Usage

# S3 method for array
reorder(x,dim=1,names=NULL,indices=NULL,FUN=mean,...)
# S3 method for matrix
reorder(x,dim=1,names=NULL,indices=NULL,FUN=mean,...)

Arguments

x

An array

dim

An integer specifying the dimension along which x should be ordered.

names

A character vector

indices

A numeric vector

FUN

A function that can be used in apply(x,dim,FUN)

...

further arguments, ignored.

Details

Typical usages are

  reorder(x,dim,names)
  reorder(x,dim,indices)
  reorder(x,dim,FUN)
  

The result of rename(x,dim,names) is x reordered such that dimnames(x)[[dim]] is equal to the concatenation of those elements of names that are in dimnames(x)[[dim]] and the remaining elements of dimnames(x)[[dim]].

The result of rename(x,dim,indices) is x reordered along dim according to indices.

The result of rename(x,dim,FUN) is x reordered along dim according to order(apply(x,dim,FUN)).

Value

The reordered object x.

See also

Examples

  (M <- matrix(rnorm(n=25),5,5,dimnames=list(LETTERS[1:5],letters[1:5])))
#>             a          b          c           d          e
#> A  0.25832927  0.1966346 -1.7460195  0.44402751  0.5773049
#> B  0.02061254  0.9553868  1.1432651 -1.12293957  0.6557535
#> C  0.09401647  0.5628267 -0.5487246  0.03841914 -0.9264028
#> D  0.78067908 -1.2167599  0.2734093 -0.71575827 -0.1826156
#> E -1.59378752  0.7186965 -1.9654935  0.13789671  1.2443059
  reorder(M,dim=1,names=c("E","A"))
#>             a          b          c           d          e
#> E -1.59378752  0.7186965 -1.9654935  0.13789671  1.2443059
#> A  0.25832927  0.1966346 -1.7460195  0.44402751  0.5773049
#> B  0.02061254  0.9553868  1.1432651 -1.12293957  0.6557535
#> C  0.09401647  0.5628267 -0.5487246  0.03841914 -0.9264028
#> D  0.78067908 -1.2167599  0.2734093 -0.71575827 -0.1826156
  reorder(M,dim=2,indices=3:1)
#>            c          b           a           d          e
#> A -1.7460195  0.1966346  0.25832927  0.44402751  0.5773049
#> B  1.1432651  0.9553868  0.02061254 -1.12293957  0.6557535
#> C -0.5487246  0.5628267  0.09401647  0.03841914 -0.9264028
#> D  0.2734093 -1.2167599  0.78067908 -0.71575827 -0.1826156
#> E -1.9654935  0.7186965 -1.59378752  0.13789671  1.2443059
  reorder(M,dim=1)
#>             a          b          c           d          e
#> E -1.59378752  0.7186965 -1.9654935  0.13789671  1.2443059
#> D  0.78067908 -1.2167599  0.2734093 -0.71575827 -0.1826156
#> C  0.09401647  0.5628267 -0.5487246  0.03841914 -0.9264028
#> A  0.25832927  0.1966346 -1.7460195  0.44402751  0.5773049
#> B  0.02061254  0.9553868  1.1432651 -1.12293957  0.6557535
  reorder(M,dim=2)
#>            c           d           a          b          e
#> A -1.7460195  0.44402751  0.25832927  0.1966346  0.5773049
#> B  1.1432651 -1.12293957  0.02061254  0.9553868  0.6557535
#> C -0.5487246  0.03841914  0.09401647  0.5628267 -0.9264028
#> D  0.2734093 -0.71575827  0.78067908 -1.2167599 -0.1826156
#> E -1.9654935  0.13789671 -1.59378752  0.7186965  1.2443059